current sql cleanup procedure moved to sqlrepository
