+++ sqlrepository/session_cleaner.go
@@ -1,4 +1,4 @@
-package cleanup
+package sqlrepository
 
 import (
 	"fmt"
